+++ CHANGELOG.md
@@ -7,3 +7,21 @@
 ## 1.0.1
 
 - Hide internal commands from command pallete
+
+## 1.0.2
+
+- Added Reality support
+- Preserve focus when execute TCL command, compile or catalog
+
+## 1.0.3
+
+- Added method Search Items (button in tree, alt+A+F or cmd+A+F)
+    Searches all open Linkar connections for item names matching the text to be searched. You will be able to open the chosen items.
+- Added method View Recent Opened Items (button in tree, alt+A+H or cmd+A+H)
+    View the last 50  opened items in any connection.
+- Added method Clear History recent opened items (alt+A+D or cmd+A+D)
+    Clear the Linkar history records
+- Added Export button in Dictionaries
+    Displays a table of all the dictionaries that can be exported to TXT or HTML.
+- Save As method
+    Now it can also be saved in another file or on another file on another open connection.